stat: Work around stack overflow in rustc
Implement parsing of /proc/<pid>/stat as a function instead of via the
`chain!` macro to limit the AST depth in expanded code. This works
around a stack overflow in some versions of rustc.

There is a performance impact, though a few hundred nanoseconds should
be acceptable in return for working on stable Rust.

    name                  old ns/iter  new ns/iter  diff ns/iter  diff %
    bench_stat            5,138        6,161               1,023  19.91%
    bench_stat_parse      2,163        2,638                 475  21.96%

Benchmarks were run with nightly-2016-10-28, a recent nightly without
the stack overflow issue.

Fixes #9
Refs https://github.com/rust-lang/rust/issues/35408